fpset.c: reimplemented without FD_SET-like masking
There is no need for FD_SET-like masking.  The comment in the code
speaks for itself: "Note that a bucket can only keep hold of such a
fingerprint that hashes into the bucket.  This obviates the need for
separate bookkeeping."

So this is what I have by the end of the year.  The first part of the
code looks good already, but the second part, which is fpset_add(),
needs more work.