fpset.c: only enable ALWAYS_GO_LEFT with FPSET_GOLEFT
The effect of the "always go left strategy" is rather small,
but the decrease in fpset_add is quite noticeable.  The decrease
in speed is especially noticeable if the strategy is activated very
early and there is a completely free bucket (so the strategy has less
effect in this case).