fp64set.h: redesigned the API/ABI
The history has been rewritten, fpset.[ch] renamed to fp64set.[ch].
This is to indicate that full 64-bit fingerprints are stored in the data
structure.  Another variant is possible, where the 64-bit fingerprint
is further digested into the smaller 32-bit fingerprint and the index.
(The match is then determined by the right fingerprint at the right
bucket; the same 32-bit fingerprint can have different meaning depending
on its location in the table.  In terms of the false positive rate,
this variant is equivalent, more or less, to the set of (32+logsize)-bit
hash values.  Thus it uses bits more optimally, but cannot deliver the
rate of false positives as low as with full 64-bit fingerprints.)

This further changes the API/ABI as follows: since the fpset_has
function call is inevitable (i.e. there is no point in inlining the
function, even if it's possible with -flto), it is best to expose vtab
which will provide the optimal implementation.  With vtab, it becomes
possible to implement various clever schemes with minimal performance
loss, such as adding a stash for elements that cannot be inserted,
or switching the structure to different number of fingerprints per
bucket: 2->3, 3->4, and 4->3 (which would yield much better memory
utilization, compared to doubling the number of buckets).