fp64set.c: commit what I have, for Christ's sake!
The bulk of the "resize" code was written on April 7th, during insomnia.
The problem is that I started to sketch it in a separate file, because
the main copy was running proba.c simulations.  Since then the code has
been subject to perfectionism-procrastination.  But this morning I can't
sleep again. :-)