fp64set.c: implemented SSE4.1-optimized has()
On Haswell, the difference is small, and I believe that non-SIMD code
may run even faster.  On Ivy Bridge, the improvement is non-negligible,
like a few percent.  I still need to think how to dispatch between the
two variants with something like IFUNC (except that I don't need
full-blown IFUNC because the functions are not exported).