fp64set.c: benchmark fpset_add return codes separately
It now shows this:

    fp64set_add[0] 31.3 x 44495
    fp64set_add[1] 157.7 x 2452
    fp64set_add[2] 17160.0 x 1
    fp64set_has 27.0 x 689116
    fp64set_free logsize=10 bsize=3 nstash=0 cnt=2453 hash=eeb7c53c1abb6b6e

Counted separately are: elements already added [0], the case similar to
fp64set_has; elements added for the first time [1], this case is about 5x
more expensive; and elements that trigger resize [2], which happens only
once.  Note that the resize takes less than 5% of add[1], which is
pretty good.  That's because in 2->3 and 3->4 resizes, no actual
reinsertion takes place.  The case 4->3 is more complex, and involves
reinsertion of about 20% of the elements (those from the last tier).
This is not too bad either.