removed FP64SET_BENCH
I found it unreliable and often misleading (on such a small function
as has() comprising only a handful of instructions).  I also tried
__rdtscp(), as opposed to plain __rdtsc(), with little luck.  Or perhaps
I'm just observing L1/L2 cache misses, which dwarf the differences
in the generated code.  But it's the code that I want to optimize.

It seems to me that the only approach left is to benchmark has()
performance in a synthetic loop:

    uint64_t t = __rdtsc();
    prngState = prngState0;
    for (int i = 0; i < (1<<20); i++) {
	bool has = fp64set_has(set, rnd());
	assert(has);
    }
    t = __rdtsc() - t;

The loop should not be too synthetic, though, in terms of memory access
patterns and branch prediction.  It would be much better to load the
loop with some real fingerprints from e.g. depfiles.c run.  This is what
in fact I'll try to do.