bench.c: benchmark fp64set_next
next2 26.4
next3 16.0
next4 12.7

It seems a bit odd at first that the price of fp64set_next drops as
the bucket size increases.  I think this is because of different fill
factor.  With bsize=2, the fill factor is 50%.  But when the table
switches to bsize=4, its fill factor with bsize=3 is about 90%,
so its fill factor after the switch is 67.5%.  This means that branch
prediction works much better.  This also means fewer cells to traverse.