fp64set-i686.S: optimize has3
-has3 20.4
+has3 17.6

I use movhps to set the high half of xmm register.  Like moveddup,
the instruction originally pertains to the floating-point domain,
and may incur a penalty on older CPUs in mixed integer/floating-point
use.  There is no penalty on modern CPUs (I run benchmarks on Ivy Bridge
and Haswell), and it actually seems to be the fastest thing to do.
gcc emits this kind of integer movhps routinely.