fp64set.h: opaque set->has asm call with reduced register clobbering
This is the technique I devised the other day.  Virtual function calls
can be made much cheaper if we instruct the compiler that the function
clobbers FEWER registers than permitted by the calling convention.
Of course, this will work well only with "primitive" functions, and
otherwise is very error-prone.  However, has() is just the case.

Note that the set argument is not modified in the routines (because I need
to access m_stash(reg_ptr) after m_bb).  This means that, in a synthetic
has() loop, there is no need to reload the argument into the register!

This also means that I'll have to reimplement the SSE2 fallback
routines, to maintain the contract.

-has2 9.92
-has3 11.09
-has4 11.31
+has2 7.02
+has3 9.08
+has4 9.03