x86: optimize insert4, likewise
This shows that only xmm6 is additionally required in insert4.  This is
only possible if one is willing to trade pblendw+pblendw+packssdw for
packssdw+packssdw+pblendw.  In reality, pblendw and packssdw have the
same cost.

This time I don't try to optimize add4 by not using xmm7, while xmm6
is strictly required.  Thus with x64 Microsoft ABI, xmm6 should already
be saved in add4 rather than in insert4.