x86: factored kick[234], reintroduced multiple branches
This works faster on all CPUs (just a little bit faster on modern Intel
microarchitectures, and a lot faster on older AMD microarchitectures).
The thing is, those branches are predictable.  For example, in kick3,
the first branch corresponding to bb[hi][0] is usually not taken because,
when the switch 2->3 takes place, the table is about 80% full; on the
other hand hand, most slots on the bb[*][2] tier are initially free.

The old code is still the best thing to do on i386, because there are
no 8-byte registers, so a comparison would take 3 instruction and two
extra registers to combine the result.

kick[234] are now separate functions, and the loop proceeds by jumping
to the beginning (so tail recursion and iteration are equivalent, much
like in Scheme).  Valgrind will now bill those functions separately.
Another peculiarity is that the number of calls valgrind accounts for
should be equivalent to the number of iterations.