removed fp64set_del and fp64set_next
For ../genbasedir/depfiles.c, I only need add() and has(); del() and
next() were developed for completeness, so that fp64set looked like a
mature data structure.

I suspect that a faster implementation of next() is possible: you can
use a separate "iterator" structure with a function pointer, and call
that instead of unified next().  Three implementations will successively
replace each other: 1) the one that traverses the leading elements with
blankValue=UINT64_MAX; 2) the bulk of the elements with blankValue=0;
3) the stash.

Also, there should be a delcur() method to delete the current element
(and update the iterator accordingly).

fp64set-x86.S is currently broken, m_* offsets need to be updated.